requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.0
decorator==4.4.2
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
frozenlist==1.6.2
fsspec==2025.5.1
greenlet==3.2.2
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moviepy==1.0.3
multidict==6.4.4
numpy==2.2.6
openai==1.84.0
orjson==3.10.18
packaging==25.0
pillow==11.2.1
proglog==0.1.12
propcache==0.3.1
psutil==7.0.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.7
safetensors==0.5.3
schedule==1.2.2
scheduler==0.8.8
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.3
typeguard==4.4.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.0
fastapi
uvicorn
sqlmodel
python-dotenv
